migrations 2/5

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -22,7 +22,7 @@ class CreateUsersTable extends Migration
             $table->timestamp('phone_verified_at')->nullable();
             $table->string('password');
             $table->uuid('default_address_id')->nullable();
-            $table->bigInteger('wallet')->default(0);
+            $table->decimal('wallet', 8, 2)->default(0);
             $table->bigInteger('reward_points')->default(0);
             // $table->rememberToken();
             $table->timestamps();
